Server/requirements.txt
Django==4.2.6
python-dotenv==1.0.0
requests==2.31.0
djangorestframework==3.14.0
openai==1.10.0
